Finding doners for charityML
Project files are for 3 ML models such as Random Forest, Gradient Boosting and  Logistic Regression etc. It explains how to interpret and use the data given for these three models. It also explains whether the data is suitable for the model designed. This project explains the use case and the supporting programming files.